first working and tested version for vectorAdd

+++ test/kernel_add.cl
@@ -1,9 +1,9 @@
 #pragma OPENCL EXTENSION cl_khr_fp64 : enable
 
 __kernel void vectorAdd(
-	__global double* bufferA,
-	__global double* bufferB,
-	__global double* result,
+	__global int* bufferA,
+	__global int* bufferB,
+	__global int* result,
 	const unsigned int size
 ){
 	int id = get_global_id(0);
